requirements.txt
streamlit==1.32.0
streamlit-webrtc==0.47.1
opencv-python==4.9.0.80
numpy==1.26.4
mediapipe==0.10.14
av==10.0.0
scikit-learn==1.3.2
google-generativeai==0.5.4